segment-validation.html

<!DOCTYPE html>
<html lang="en">
  <head>
    <meta charset="utf-8" />
    <title>Segment Validation</title>
    <link rel="stylesheet" href="../lib/style.css" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>

  <body>
    <div id="content-wrapper">
      <div id="video-wrapper">
        <video></video>
        <div><div id="qr-code"></div></div>
        <div><div id="status-text"></div></div>
      </div>
      <div>To show/hide debug overlay press up/down or button beneath</div>
      <div id="debug-wrapper">
        <div>
          <div class="button" id="debug-button">Debug Overlay</div>
        </div>
        <div id="debug"></div>
        <div id="log"></div>
      </div>
    </div>
    <div id="info-overlay"></div>
    <script src="../lib/mozilla/object-keys-polyfill.js"></script>
    <script src="../lib/stefanpenner/es6-promise.min.js"></script>
    <script src="../lib/player.js"></script>
    <script src="../lib/manifest_parser.js"></script>
    <script src="../lib/mpd-parser.js"></script>
    <script src="../lib/wave-service.js"></script>
    <script src="../lib/davidshimjs/qrcode.js"></script>
    <script src="../lib/dpctf-testharness.js"></script>
    <script src="../lib/hbbtv.js"></script>
    <script src="../lib/segmentValidation/mp4box.all.global.js"></script>
    <script src="../lib/segmentValidation/segmentAnalyzer.js"></script>
    <script>
      
      // Global variables
      var TEST_INFO = {
        id: "segment-validation",
        title: "Segment Validation",
        description:
          "Test if segments exists on server side with HTTP HEAD requests.",
        code: "{{TEMPLATE_NAME}}",
      };

      var video = document.querySelector("video");
      var qrCode = document.getElementById("qr-code");
      var statusText = document.getElementById("status-text");

      var videoContentModel = "{{VIDEO_MPD_URL}}";
      var audioContentModel = "{{AUDIO_MPD_URL}}";

      var dpctfTest = new DpctfTest({
        testInfo: TEST_INFO,
        videoContentModel: videoContentModel,
        audioContentModel: audioContentModel,
        videoElement: video,
        qrCodeElement: qrCode,
        statusTextElement: statusText,
        infoOverlayElement: document.getElementById("info-overlay"),
        executeTest: executeTest,
        usePlayout: false,
      });

      async function executeTest(player, done, parameters) {
        var logger = dpctfTest.getLogger();
        var minBufferDuration = parameters.minBufferDuration / 1000;
        player.startBuffering();

        // save MPD file as utf-8 string
        let mpd = await fetch(videoContentModel[0])
        .then(res => res.text());
        mpd = JSON.parse( JSON.stringify( mpd ) );
        logger.log("MPD loaded successfully");
        
        let baseUrl = videoContentModel[0].split("/").filter(e => !e.includes(".mpd")).join("/") + "/";
        let segmentList = mpdParser.parse(mpd);

        // adjust extracted uris that they can be requested
        let segmentUri = segmentList.playlists[0].segments.map(e => e.uri).map(e => baseUrl + e);
        let segmentDuration = segmentList.playlists[0].segments.map(e => e.duration);

        // test case that fetch all segments if they exist on server side
        dpctfTest.asyncTest(function (test) {
          var query = location.search.replace(/\?/, "");
          var match = query.match(/redirect_time=([^&]+)/);
          var REDIRECT_TIME = match ? match[1] : null;
          if (!REDIRECT_TIME) REDIRECT_TIME = 5;

          let testPass = true;
          let segmentTest = [];
          for (let i = 0; i < segmentUri.length; i++) {
            segmentTest.push(fetch(segmentUri[i], {method: "HEAD"})
            .then(res => {
                if (String(res.status).startsWith("4") || String(res.status).startsWith("5")) {
                  logger.log(`Segment ${i} not exist! Test failed.`);
                  testPass = false;
                } else {
                  logger.log(`Segment ${i} exist.`);
                }
            }));
          }

          Promise.all(segmentTest)
          .then(() => {
            test.step(() => assert_true(testPass, "Segments should exist on server side"));
            test.done();
          })
          .catch(err => {
            logger.log(err.message);
            console.error(err);
            test.step(() => assert_false(false, err));
            test.done();
          });

        }, "Segment exist on Server");

        // test case that compare the segment durations with durations from the MPD
        dpctfTest.asyncTest(async function (test) {
          var query = location.search.replace(/\?/, "");
          var match = query.match(/redirect_time=([^&]+)/);
          var REDIRECT_TIME = match ? match[1] : null;
          if (!REDIRECT_TIME) REDIRECT_TIME = 5;

          let resultArr = await startAnalyzing();
          test.step(() => assert_true(resultArr.every(results => results == true), "All segment durations should fit with the MPD"));
          test.done();
        }, "Segment duration correct");

        async function startAnalyzing() {
          // Fetch the init file
          let initUrl = segmentUri[0].split("/").filter(e => !e.includes('.m4s')).join('/') + "/init.mp4";
          let init = await fetch(initUrl)
          .then(res => res.arrayBuffer());

          // Analyze every segment parallel
          let results = [];
          for (let i = 0; i < segmentUri.length; i++) {
              results.push(
                fetch(segmentUri[i])
                .then(res => res.arrayBuffer())
                .then(segment => segmentAnalyzer.analyzeSegment(init, segment, segmentDuration[i], logger, i))
            );               
          }
          return await Promise.all(results).then(res => res.flat());
        }
        
        player.playOnBufferLoaded(minBufferDuration).then(done);
      }
    </script>
  </body>
</html>
